Fixed int problem

+++ src/transformer_with_adapters.py
@@ -262,7 +262,7 @@ class TransformerWithAdapters:
             if self.do_query:
                 samples_entropy = torch.topk(samples_entropy_all, self.query_samples_count)
             else:
-                samples_entropy = torch.topk(samples_entropy_all, unlabeled_dataset.num_rows*self.query_samples_ratio)
+                samples_entropy = torch.topk(samples_entropy_all, int(unlabeled_dataset.num_rows*self.query_samples_ratio))
 
             new_train_samples = unlabeled_dataset.select(samples_entropy.indices.tolist())
